fixValueMax

+++ src/main/java/com/itmo/java/basics/logic/io/DatabaseInputStream.java
@@ -30,17 +30,15 @@ public class DatabaseInputStream extends DataInputStream {
         try {
             Optional<DatabaseRecord> rec;
 
-            byte keySize = readByte();
-            //int keySize = readInt();
+            //byte keySize = readByte();
+            int keySize = readInt();
             String keyString = new String(readNBytes(keySize), StandardCharsets.UTF_8);
-            //String keyString = readUTF();
 
-            byte valueSize = readByte();
-            //int valueSize = readInt();
+            //byte valueSize = readByte();
+            int valueSize = readInt();
             if(valueSize == -1)
                 return Optional.empty();
             String valueString = new String(readNBytes(valueSize), StandardCharsets.UTF_8);
-            // String valueString = readUTF();
 
             rec = Optional.of(SetDatabaseRecord.create(keyString, valueString.getBytes(StandardCharsets.UTF_8)));
